% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_ASCVD-PCE_func.R
\name{ASCVD}
\alias{ASCVD}
\title{ASCVD (Atherosclerotic Cardiovascular Disease) Risk Algorithm including Known ASCVD from AHA/ACC}
\usage{
ASCVD(
  Gender = Gender,
  Ethnicity = Ethnicity,
  Age = Age,
  total.chol = total.chol,
  total.hdl = total.hdl,
  systolic.bp = systolic.bp,
  hypertension = hypertension,
  smoker = smoker,
  diabetes = diabetes,
  classify = FALSE
)
}
\arguments{
\item{Gender}{a binary character vector of sex values. Categories should include only 'male' or 'female'}

\item{Ethnicity}{a character vector, 'white', 'black', 'asian', or other}

\item{Age}{a numeric vector of age values, in years}

\item{total.chol}{a numeric vector of total cholesterol values, in mmol/L}

\item{total.hdl}{a numeric vector of total high density lipoprotein HDL values, in mmol/L}

\item{systolic.bp}{a numeric vector of systolic blood pressure continuous values}

\item{hypertension}{a binary numeric vector, 1 = yes and 0 = no}

\item{smoker}{a binary numeric vector, 1 = yes and 0 = no}

\item{diabetes}{a binary numeric vector, 1 = yes and 0 = no}

\item{classify}{a logical parameter to indicate classification of Scores "TRUE" or none "FALSE"}
}
\value{
A vector with ASCVD score calculations
and/or a vector of their classifications if indicated
}
\description{
This function implements the ASCVD score calculation as a vector

Scored using 3 steps

Step 1: High-Risk Criteria: History of ASCVD
History of ASCVD - History of acute coronary syndrome (ACS),
myocardial infarction (MI), stable angina,
coronary/other arterial revascularization,
stroke, transient ischemic attack,
or Peripheral Arterial Disease (PAD) from atherosclerosis

Step 2: High-Risk Criteria: Extreme LDL
LDL Cholesterol ≥190mg/dL (4.92 mmol/L)

Step 3: ASCVD Risk Criteria; Only Apply When LDL 70-189mg/dL (1.81-4.90 mmol/L)
Age
Diabetes
Total cholesterol
HDL cholesterol
Systolic BP
Treatment for Hypertension
Smoker
Race
Balck british as Black; everything else as White

Three possible outcome

High-Intensity Statin Therapy
Moderate-Intensity Statin Therapy
Low-Intensity Statin Therapy
}
\examples{
results <- cohort_xx \%>\% rowwise() \%>\%
 mutate(ASCVD_score = ASCVD(Gender, Ethnicity, Age, total.chol, total.hdl,
 systolic.bp,hypertension, smoker, diabetes, classify = FALSE))

}
\keyword{ASCVD,}
\keyword{Age,}
\keyword{Ethnicity,}
\keyword{Gender,}
\keyword{classify}
\keyword{diabetes,}
\keyword{smoker,}
\keyword{systolic.bp,hypertension,}
\keyword{total.chol,}
\keyword{total.hdl,}
