% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_EDACS_func.R
\name{EDACS}
\alias{EDACS}
\title{Emergency Department Assessment of Chest Pain Score (EDACS)}
\usage{
EDACS(
  Age,
  Gender,
  diabetes,
  smoker,
  hypertension,
  hyperlipidaemia,
  family.history,
  sweating,
  pain.radiation,
  pleuritic,
  palpation,
  ecg.st.depression,
  ecg.twi,
  presentation_hstni,
  Δ2nd_hstni,
  classify
)
}
\arguments{
\item{Age}{a numeric vector of age values, in years}

\item{Gender}{a binary character vector of sex values. Categories should include
only 'male' or 'female'.}

\item{diabetes}{a binary numeric vector, 1 = yes and 0 = no}

\item{smoker}{a binary numeric vector, 1 = yes and 0 = no}

\item{hypertension}{a binary numeric vector, 1 = yes and 0 = no}

\item{hyperlipidaemia}{a binary numeric vector, 1 = yes and 0 = no}

\item{family.history}{a binary numeric vector, 1 = yes and 0 = no}

\item{sweating}{a binary numeric vector, 1 = yes and 0 = no}

\item{pain.radiation}{a binary numeric vector, 1 = yes and 0 = no}

\item{pleuritic}{a binary numeric vector, 1 = yes and 0 = no}

\item{palpation}{a binary numeric vector, 1 = yes and 0 = no}

\item{ecg.st.depression}{a binary numeric vector, 1 = yes and 0 = no}

\item{ecg.twi}{a binary numeric vector, 1 = yes and 0 = no}

\item{presentation_hstni}{a continuous numeric vector of the troponin levels}

\item{typical_symptoms.num}{a numeric vector of the number of typical symptoms}

\item{ecg.normal}{a binary numeric vector,  1 = yes and 0 = no}

\item{abn.repolarisation}{a binary numeric vector,  1 = yes and 0 = no}

\item{atherosclerotic.disease}{a binary numeric vector, 1 = yes and 0 = no}

\item{2nd_hstni}{a binary numeric vector, 1 = yes and 0 = no}
}
\value{
A vector with EDACS score calculations
and/or a vector of their classifications if indicated
}
\description{
This function implements the EDACS score calculation as a vector

Needed variables -------------------------------------------------------
Age -
 18-45 = 2
46-50 = 4
51-55 = 6     All available
56-60 = 8
61-65 = 10
66-70 = 12
71-75 = 14
76-80 = 16
81-85 = 18
≥86 = 20

Sex -
Female = 0    All available
Male = 6

Known coronary artery disease or ≥3 risk factors*  The risk factors only apply to patients 18-50-
no = 0                     All available
yes = 4

Symptoms and signs
Diaphoresis                                        no = 0   yes = 3
Pain radiates to arm, shoulder, neck, or jaw       no = 0  yes = 5
Pain occurred or worsened with inspiration    NA   no = 0  yes = -4
Pain is reproduced by palpation               NA   no = 0  yes = -6

Possible outcomes -------------------------------------------------------

Two possible outcomes

Low risk cohort:
EDACS <16 and
EKG shows no new ischemia and
0-hr and 2-hr troponin both negative.
Not low risk cohort:
EDACS ≥16 or
EKG shows new ischemia or
0-hr or 2-hr troponin positive.
}
\examples{
 results <- cohort_xx \%>\% rowwise() \%>\% mutate(EDACS_score = EDACS(Age, Gender, diabetes, smoker, hypertension, hyperlipidaemia, family.history, sweating, pain.radiation, pleuritic, palpation, ecg.st.depression, ecg.twi,  presentation_hstni, Δ2nd_hstni, classify = FALSE))

}
\keyword{Age,}
\keyword{EDACS,}
\keyword{Gender,}
\keyword{classify}
\keyword{diabetes,}
\keyword{ecg.st.depression,}
\keyword{ecg.twi,}
\keyword{family.history,}
\keyword{hyperlipidaemia,}
\keyword{hypertension,}
\keyword{pain.radiation,}
\keyword{palpation,}
\keyword{pleuritic,}
\keyword{presentation_hstni,}
\keyword{smoker,}
\keyword{sweating,}
\keyword{Δ2nd_hstni,}
