% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_GRACE_df.R
\name{GRACE_scores}
\alias{GRACE_scores}
\title{GRACE Global Registry of Acute Coronary Events version 2.0, 6 months outcome}
\usage{
GRACE_scores(data, classify)
}
\arguments{
\item{data}{A data frame with all the variables needed for calculation:
killip.class, systolic.bp, heart.rate, Age, creat, ecg.st.depression,
presentation_hstni, cardiac.arrest, Gender, classify}
}
\value{
data frame with two extra columns including the 'GRACE_score' calculations
and their classifications, 'GRACE_strat'
}
\description{
This function allows you to calculate the GRACE 2.0 score row wise
in a data frame with the required variables. It would then retrieve a data
frame with two extra columns including the calculations and their classifications
}
\examples{

# Create a data frame or list with the necessary variables
# Set the number of rows
num_rows <- 100
# Create a larger dataset with 100 rows
cohort_xx <- data.frame(
  typical_symptoms.num = as.numeric(sample(0:6, num_rows, replace = TRUE)),
  ecg.normal = as.numeric(sample(c(0, 1), num_rows, replace = TRUE)),
  abn.repolarisation = as.numeric(sample(c(0, 1), num_rows, replace = TRUE)),
  ecg.st.depression = as.numeric(sample(c(0, 1), num_rows, replace = TRUE)),
  Age = as.numeric(sample(30:80, num_rows, replace = TRUE)),
  diabetes = sample(c(1, 0), num_rows, replace = TRUE),
  smoker = sample(c(1, 0), num_rows, replace = TRUE),
  hypertension = sample(c(1, 0), num_rows, replace = TRUE),
  hyperlipidaemia = sample(c(1, 0), num_rows, replace = TRUE),
  family_history = sample(c(1, 0), num_rows, replace = TRUE),
  atherosclerotic.disease = sample(c(1, 0), num_rows, replace = TRUE),
  presentation_hstni = as.numeric(sample(10:100, num_rows, replace = TRUE)),
  Gender = sample(c("male", "female"), num_rows, replace = TRUE),
  sweating = as.numeric(sample(c(0, 1), num_rows, replace = TRUE)),
  pain.radiation = as.numeric(sample(c(0, 1), num_rows, replace = TRUE)),
  pleuritic = as.numeric(sample(c(0, 1), num_rows, replace = TRUE)),
  palpation = as.numeric(sample(c(0, 1), num_rows, replace = TRUE)),
  ecg.twi = as.numeric(sample(c(0, 1), num_rows, replace = TRUE)),
  Δ2nd_hstni = as.numeric(sample(1:200, num_rows, replace = TRUE)),
  killip.class = as.numeric(sample(1:4, num_rows, replace = TRUE)),
  systolic.bp = as.numeric(sample(0:300, num_rows, replace = TRUE)),
  heart.rate = as.numeric(sample(0:300, num_rows, replace = TRUE)),
  creat = as.numeric(sample(0:4, num_rows, replace = TRUE)),
  cardiac.arrest = as.numeric(sample(c(0, 1), num_rows, replace = TRUE))
)
# Call the function with the cohort_xx
result <- GRACE_scores(data = cohort_xx, classify = TRUE)
Print the results
summary(result$GRACE_strat)
summary(result$GRACE_score)

}
\keyword{Age,}
\keyword{GRACE,}
\keyword{Gender,}
\keyword{cardiac.arrest,}
\keyword{classify}
\keyword{creat,}
\keyword{ecg.st.depression,}
\keyword{heart.rate,}
\keyword{killip.class,}
\keyword{presentation_hstni,}
\keyword{systolic.bp,}
