% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rounding.R
\name{round_to_nearest_digit}
\alias{round_to_nearest_digit}
\title{Round To Nearest Digit}
\usage{
round_to_nearest_digit(number, digits = 0)
}
\arguments{
\item{number}{Input numeric value}

\item{digits}{Input integer indicating the number of
decimal places to be used. By default, it symmetrically rounds
off to the nearest integer.
Default: 0}
}
\description{
A function that does
symmetric rounding to the nearest digits.
}
\examples{
round_to_nearest_digit(0.5)
round_to_nearest_digit(1.5)
round_to_nearest_digit(-0.5)
round_to_nearest_digit(-1.5)
}
