% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_SCORE2_func.R
\name{SCORE2/OP}
\alias{SCORE2/OP}
\alias{SCORE2}
\title{Systematic COronary Risk Evaluation (SCORE) model}
\usage{
SCORE2(
  Age,
  Gender,
  smoker,
  systolic.bp,
  diabetes,
  total.chol,
  total.hdl,
  classify
)
}
\arguments{
\item{Age}{a numeric vector of age values, in years}

\item{Gender}{a binary character vector of Gender values. Categories should include
only 'male' or 'female'.}

\item{smoker}{a binary numeric vector, 1 = yes and 0 = no}

\item{systolic.bp}{a numeric vector of systolic blood pressure continuous values}

\item{total.chol}{a numeric vector of total cholesterol values, in mmol/L}

\item{total.total.hdl}{a numeric vector of total high density lipoprotein total.hdl values, in mmol/L}
}
\value{
A vector with SCORE2/OP score calculations
and/or a vector of their classifications if indicated
}
\description{
This function implements the ASCVD score calculation as a vector

formula in SCORE2 Updated Supplementary Material page  9. paper:
"SCORE2 risk prediction algorithms: new models to estimate 10-year risk of cardiovascular disease in Europe"

Age	10-year risk of fatal and non-fatal cardiovascular disease
              Low risk    Moderate risk     High risk
< 50 years   	  <2.5%	  |   2.5 to <7.5%  | 	≥7.5%
50 - 69 years	  <5%     |   5 to <10%	    |   ≥10%
 ≥ 70 years	    <7.5%	  |  7.5 to <15% 	  |   ≥15%

above classifications referred from https://www.inanutshell.ch/en/digital-doctors-bag/score2-and-score2-op/#:~:text=SCORE2%20(%C2%ABSystematic%20COronary%20Risk%20Evaluation,Society%20of%20Cardiology%20(ESC).
}
\examples{
  results <- cohort_xx \%>\% rowwise() \%>\% mutate(SCORE2OP_score = SCORE2(Age, Gender, smoker, systolic.bp, diabetes, total.chol, total.hdl, classify = FALSE))

}
\keyword{Age,}
\keyword{Gender,}
\keyword{SCORE2/OP,}
\keyword{classify}
\keyword{diabetes,}
\keyword{smoker,}
\keyword{systolic.bp,}
\keyword{total.chol,}
\keyword{total.hdl,}
