% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_calc_all_scores_prec.R
\name{RiskScoresCalc}
\alias{RiskScoresCalc}
\alias{calc_scores}
\title{Commonly used cardiovascular risk scores for the prediction of major cardiac events (MACE)}
\usage{
calc_scores(
  data,
  typical_symptoms.num_df = "typical_symptoms.num",
  ecg.normal_df = "ecg.normal",
  abn.repolarisation_df = "abn.repolarisation",
  ecg.st.depression_df = "ecg.st.depression",
  Age_df = "Age",
  diabetes_df = "diabetes",
  smoker_df = "smoker",
  hypertension_df = "hypertension",
  hyperlipidaemia_df = "hyperlipidaemia",
  family.history_df = "family.history",
  atherosclerotic.disease_df = "atherosclerotic.disease",
  presentation_hstni_df = "presentation_hstni",
  Gender_df = "Gender",
  sweating_df = "sweating",
  pain.radiation_df = "pain.radiation",
  pleuritic_df = "pleuritic",
  palpation_df = "palpation",
  ecg.twi_df = "ecg.twi",
  Δ2nd_hstni_df = "Δ2nd_hstni",
  killip.class_df = "killip.class",
  heart.rate_df = "heart.rate",
  systolic.bp_df = "systolic.bp",
  aspirin_df = "aspirin",
  number.of.episodes.24h_df = "number.of.episodes.24h",
  previous.pci_df = "previous.pci",
  previous.cabg_df = "previous.cabg",
  total.chol_df = "total.chol",
  total.hdl_df = "total.hdl",
  Ethnicity_df = "Ethnicity"
)
}
\arguments{
\item{typical_symptoms.num}{a numeric vector of the number of typical symptoms}

\item{ecg.normal}{a binary numeric vector,  1 = yes and 0 = no}

\item{abn.repolarisation}{a binary numeric vector,  1 = yes and 0 = no}

\item{ecg.st.depression}{a binary numeric vector, 1 = yes and 0 = no}

\item{Age}{a numeric vector of age values, in years}

\item{diabetes}{a binary numeric vector, 1 = yes and 0 = no}

\item{smoker}{a binary numeric vector, 1 = yes and 0 = no}

\item{hypertension}{a binary numeric vector, 1 = yes and 0 = no}

\item{hyperlipidaemia}{a binary numeric vector, 1 = yes and 0 = no}

\item{family.history}{a binary numeric vector, 1 = yes and 0 = no}

\item{atherosclerotic.disease}{a binary numeric vector, 1 = yes and 0 = no}

\item{presentation_hstni}{a continuous numeric vector of the troponin levels}

\item{Gender}{a binary character vector of sex values. Categories should include
only 'male' or 'female'.}

\item{sweating}{a binary numeric vector, 1 = yes and 0 = no}

\item{pain.radiation}{a binary numeric vector, 1 = yes and 0 = no}

\item{pleuritic}{a binary numeric vector, 1 = yes and 0 = no}

\item{palpation}{a binary numeric vector, 1 = yes and 0 = no}

\item{ecg.twi}{a binary numeric vector, 1 = yes and 0 = no}

\item{2nd_hstni}{a binary numeric vector, 1 = yes and 0 = no}

\item{killip.class}{a numeric vector of killip class values, 1 to 4}

\item{heart.rate}{a numeric vector of heart rate continuous values}

\item{systolic.bp}{a numeric vector of systolic blood pressure continuous values}

\item{aspirin}{a binary numeric vector, 1 = yes and 0 = no}

\item{number.of.episodes.24h}{a numeric vector of number of angina episodes in 24 hours}

\item{previous.pci}{a binary numeric vector, 1 = yes and 0 = no}

\item{previous.cabg}{a binary numeric vector, 1 = yes and 0 = no}

\item{total.chol}{a numeric vector of total cholesterol values, in mmol/L}

\item{total.hdl}{a numeric vector of total high density lipoprotein HDL values, in mmol/L}
}
\value{
a data frame with two extra columns including all the cardiovascular risk score calculations
and their classifications
}
\description{
This function implements seven cardiovascular risk scores row wise
in a data frame with the required variables. It would then retrieve a data
frame with two extra columns for each risk score including their calculations and classifications
}
\examples{
new_data_frame <- calc_scores(data = cohort_xx)

}
\keyword{2nd_hstni,}
\keyword{Age,}
\keyword{Cardiovascular}
\keyword{Gender,}
\keyword{abn.repolarisation,}
\keyword{aspirin,}
\keyword{atherosclerotic.disease,}
\keyword{diabetes,}
\keyword{ecg.normal,}
\keyword{ecg.st.depression,}
\keyword{ecg.twi,}
\keyword{family.history,}
\keyword{heart.rate,}
\keyword{hyperlipidaemia,}
\keyword{hypertension,}
\keyword{killip.class,}
\keyword{number.of.episodes.24h,}
\keyword{pain.radiation,}
\keyword{palpation,}
\keyword{pleuritic,}
\keyword{presentation_hstni,}
\keyword{previous.cabg,}
\keyword{previous.pci,}
\keyword{risk}
\keyword{scores,}
\keyword{smoker,}
\keyword{sweating,}
\keyword{systolic.bp,}
\keyword{total.chol,}
\keyword{total.hdl}
\keyword{typical_symptoms.num,}
